gis_test/exercise_1/requirements.txt
python-crontab==1.9.3
requests[security]==2.7
psycopg2==2.6.1
postgres==2.1.1